requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic>=2.0,<3.0
pydantic-settings>=2.12

langchain==0.3.3
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
pdfplumber==0.9.0
python-docx==0.8.11
openai>=1.40.0

sqlalchemy>=2.0,<3.0
psycopg[binary]==3.2.1
pgvector==0.3.0

alembic==1.13.1

python-dotenv==1.0.1
pypandoc==1.13
